Splitting routes file
